feat: Line 지하철 노선 생성시 이름에 대한 예외 검사 기능 구현
- 지하철 노선인 Line 객체는 이름이 null, 공백이 아닌 2글자 이상이다.
- 예외인 경우, LineNameException 커스텀 예외를 호출한다.
- 클래스 위치를 entity 패키지 내부로 이동시킨다.
