+++ src/main/java/subway/domain/entity/Line.java
@@ -1,0 +1,28 @@
+package subway.domain.entity;
+
+import java.util.Objects;
+
+public class Line {
+    private static final int MINIMUM_LINE_NAME_LENGTH = 2;
+
+    private final String name;
+
+    public Line(String name) {
+        validateName(name);
+        this.name = name;
+    }
+
+    private void validateName(String name) {
+        if (Objects.isNull(name)) {
+            throw new LineNameException();
+        }
+        int trimNameLength = name.trim().length();
+        if (trimNameLength < MINIMUM_LINE_NAME_LENGTH) {
+            throw new LineNameException();
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+}
